apps/backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
slowapi==0.1.9
psycopg2-binary==2.9.11
bcrypt==5.0.0
pyjwt==2.10.1
requests==2.32.5
google-cloud-texttospeech==2.34.0
alembic==1.18.1
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
